refactor and add cron job to sync PR to jira
