+++ sync_issues_to_jira/Dockerfile
@@ -14,6 +14,7 @@ RUN apt-get install -y curl && curl -sL https://deb.nodesource.com/setup_11.x | 
 RUN apt-get install -y nodejs npm && npm i markdown2confluence -g
 
 ADD sync_issue.py /sync_issue.py
-ADD test_sync_issue.py /test_sync_issue.py
+ADD sync_pr.py /sync_pr.py
+ADD sync_to_jira.py /sync_to_jira.py
 
-ENTRYPOINT ["/sync_issue.py"]
+ENTRYPOINT ["/sync_to_jira.py"]
